completed exercises 4.15-4.23

+++ part4/controllers/blogs.js
@@ -1,18 +1,37 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const mongoose = require("mongoose");
+const { userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   res.json(blogs);
 });
 
-blogsRouter.post("/", async (req, res) => {
-  const savedBlog = await new Blog(req.body).save();
+blogsRouter.post("/", userExtractor, async (req, res) => {
+  const { title, author, url, likes } = req.body;
+  const user = req.user;
+
+  const savedBlog = await new Blog({
+    title,
+    author,
+    url,
+    likes,
+    user: user._id,
+  }).save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   res.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (req, res) => {
+blogsRouter.delete("/:id", userExtractor, async (req, res) => {
+  const user = req.user;
+  const blog = await Blog.findById(req.params.id);
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return res
+      .status(401)
+      .json({ error: "you are not authorized to delete this resource" });
+  }
   await Blog.findByIdAndRemove(req.params.id);
   res.status(204).end();
 });
@@ -20,7 +39,6 @@ blogsRouter.delete("/:id", async (req, res) => {
 blogsRouter.put("/:id", async (req, res) => {
   const { title, author, url, likes } = req.body;
 
-  // Wasn't able to get the model validation working with mongoose -kept returning 200
   if (!title || !author) {
     return res.status(400).end();
   }
@@ -33,7 +51,6 @@ blogsRouter.put("/:id", async (req, res) => {
       context: "query",
     }
   );
-  console.log(updatedBlog);
   res.status(200).json(updatedBlog);
 });
